added config in c-make

+++ templates/c_make/create_make_root.sh
@@ -7,5 +7,11 @@ if [[ -f "Makefile" ]]; then
     exit 1
 fi
 
+if [[ -f ".config.mk" ]]; then
+    echo "There already exists a config file: Will not replace it."
+    exit 2
+fi
+
 # === Copy the Makefile ===
 cp $make_root_temp "Makefile"
+cp $conf_temp ".config.mk"
